criacao de testes para a cara editoracontroller, validacoes adicionadas na classe de editora controle

+++ BibliotecaController/src/br/com/anhanguera/pos/biblioteca/controller/EditoraController.java
@@ -11,6 +11,7 @@ import java.util.List;
 public class EditoraController {
     
     private Editora editora;
+    public static String msgErro = "";
     
     public EditoraController(Editora editora){
         this.editora = editora;
@@ -20,7 +21,12 @@ public class EditoraController {
     
     public boolean insert(){
         try{
-            return new EditoraDAO().insert(this.editora);
+            if(this.editora.getNomeEditora().equals(null))
+                return false;
+            else if(this.editora.getCidadeEditora().equals(null))
+                return false;
+            else
+                return new EditoraDAO().insert(this.editora);
         }catch(Exception e){
             return false;
         }
@@ -28,7 +34,17 @@ public class EditoraController {
     
     public boolean alter(){
         try{
-            return new EditoraDAO().alter(this.editora);
+            if(!new EditoraDAO().exist(this.editora.getCodigoEditora()))
+                return false;
+            
+            if(this.editora.getCodigoEditora() == 0)
+                return false;
+            else if (this.editora.getNomeEditora().equals(null))
+                return false;
+            else if (this.editora.getCidadeEditora().equals(null))
+                return false;
+            else
+                return new EditoraDAO().alter(this.editora);
         }catch(Exception e){
             return false;
         }
@@ -36,7 +52,14 @@ public class EditoraController {
     
     public boolean delete(int codigoEditora){
         try{
-            return new EditoraDAO().delete(codigoEditora);
+            
+            if(!new EditoraDAO().exist(codigoEditora))
+                return false;
+            
+            if(codigoEditora == 0)
+                return false;
+            else
+                return new EditoraDAO().delete(codigoEditora);
         }catch(Exception e){
             return false;
         }
@@ -56,5 +79,10 @@ public class EditoraController {
         }catch(Exception e){
             return null;
         }
+    }
+    
+    public void TruncaTable()
+    {
+        new EditoraDAO().TruncaTable();
     }
 }
